+++ workshop/src/index.ts
@@ -4,6 +4,7 @@ import dixtPluginJoin from "dixt-plugin-join";
 import dixtPluginLogs from "dixt-plugin-logs";
 import dixtPluginPresence from "dixt-plugin-presence";
 import dixtPluginReact from "dixt-plugin-react";
+import dixtPluginReports from "dixt-plugin-reports";
 import dixtPluginRoles from "dixt-plugin-roles";
 import dixtPluginTwitch from "dixt-plugin-twitch";
 import dixtPluginWorktime from "dixt-plugin-worktime";
@@ -11,6 +12,7 @@ import dixtPluginWorktime from "dixt-plugin-worktime";
 import dixtPluginAffixOptions from "./options/affix";
 import dixtPluginJoinOptions from "./options/join";
 import dixtPluginReactOptions from "./options/react";
+import dixtPluginReportsOptions from "./options/reports";
 import dixtPluginRolesOptions from "./options/roles";
 import dixtPluginTwitchOptions from "./options/twitch";
 import dixtPluginWorktimeOptions from "./options/worktime";
@@ -27,6 +29,7 @@ const main = async () => {
       [dixtPluginAffix, dixtPluginAffixOptions],
       [dixtPluginJoin, dixtPluginJoinOptions],
       [dixtPluginReact, dixtPluginReactOptions],
+      [dixtPluginReports, dixtPluginReportsOptions],
       [dixtPluginTwitch, dixtPluginTwitchOptions],
       [dixtPluginWorktime, dixtPluginWorktimeOptions],
       [dixtPluginRoles, dixtPluginRolesOptions],