+++ workshop/src/options/reports.ts
@@ -1,0 +1,9 @@
+import { DixtPluginReportsOptions } from "dixt-plugin-reports";
+
+import { CHANNELS } from "../constants";
+
+const dixtPluginReportsOptions: DixtPluginReportsOptions = {
+  channel: CHANNELS.DIXT_PLUGIN_REPORTS.REPORTS,
+};
+
+export default dixtPluginReportsOptions;