+++ workshop/src/options/twitch.ts
@@ -1,10 +1,10 @@
 import { DixtPluginTwitchOptions } from "dixt-plugin-twitch";
 
-import { CHANNELS } from "../constants";
+import { CHANNELS, ROLES } from "../constants";
 
 const dixtPluginTwitchOptions: DixtPluginTwitchOptions = {
   channel: CHANNELS.DIXT_PLUGIN_TWITCH.STREAMS,
-  roles: ["1095280448072069180"],
+  roles: [ROLES.MAINTAINER, ROLES.CONTRIBUTOR],
 };
 
 export default dixtPluginTwitchOptions;