+++ packages/dixt-plugin-affix/CHANGELOG.md
@@ -1,5 +1,19 @@
 # dixt-plugin-affix
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [3cdb21f]
+  - dixt@2.1.8
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [c91ac30]
+  - dixt@2.1.7
+
 ## 1.0.4
 
 ### Patch Changes