+++ packages/dixt-plugin-join/CHANGELOG.md
@@ -1,5 +1,19 @@
 # dixt-plugin-join
 
+## 2.1.10
+
+### Patch Changes
+
+- Updated dependencies [3cdb21f]
+  - dixt@2.1.8
+
+## 2.1.9
+
+### Patch Changes
+
+- Updated dependencies [c91ac30]
+  - dixt@2.1.7
+
 ## 2.1.8
 
 ### Patch Changes