+++ packages/dixt-plugin-react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # dixt-plugin-react
 
+## 2.1.8
+
+### Patch Changes
+
+- Updated dependencies [3cdb21f]
+  - dixt@2.1.8
+
+## 2.1.7
+
+### Patch Changes
+
+- Updated dependencies [c91ac30]
+  - dixt@2.1.7
+
 ## 2.1.6
 
 ### Patch Changes