+++ packages/dixt-plugin-roles/CHANGELOG.md
@@ -1,5 +1,20 @@
 # dixt-plugin-roles
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [3cdb21f]
+  - dixt@2.1.8
+
+## 1.0.2
+
+### Patch Changes
+
+- c91ac30: Add discord utils
+- Updated dependencies [c91ac30]
+  - dixt@2.1.7
+
 ## 1.0.1
 
 ### Patch Changes