+++ packages/dixt/CHANGELOG.md
@@ -1,5 +1,17 @@
 # dixt
 
+## 2.1.8
+
+### Patch Changes
+
+- 3cdb21f: Remove commands support until it isnt ready
+
+## 2.1.7
+
+### Patch Changes
+
+- c91ac30: Add discord utils
+
 ## 2.1.6
 
 ### Patch Changes